+++ test/unit/pt_osc_migration_unit_test.rb
@@ -56,7 +56,7 @@ class PtOscMigrationUnitTest < Test::Unit::TestCase
 
           should 'set missing flags to default values' do
             flags_with_defaults = ActiveRecord::PtOscMigration.percona_flags.select do |flag, config|
-              config.key?(:default) && flag != 'execute' && !config[:boolean]
+              !config[:default].nil? && flag != 'execute' && !config[:boolean]
             end
 
             command = @migration.send(:percona_command, '', '', '')
@@ -161,6 +161,98 @@ class PtOscMigrationUnitTest < Test::Unit::TestCase
             @migration.send(:percona_command, '', '', '', @options)
           end
         end
+
+        context 'user flag' do
+          context 'with normal database username specified' do
+            setup do
+              @migration.stubs(:raw_database_config).returns(username: 'foobar')
+            end
+
+            should 'add the user to the command' do
+              command = @migration.send(:percona_command, '', '', '')
+              assert command.include?('--user foobar'), "command #{command} did not include user"
+            end
+          end
+
+          context 'with user specified in percona config' do
+            setup do
+              @migration.stubs(:raw_database_config).returns(percona: { user: 'foobar' })
+            end
+
+            should 'add the user to the command' do
+              command = @migration.send(:percona_command, '', '', '')
+              assert command.include?('--user foobar'), "command #{command} did not include user"
+            end
+          end
+
+          context 'with user specified in database and percona config' do
+            setup do
+              @migration.stubs(:raw_database_config).returns(username: 'foobar', percona: { user: 'bazqux' })
+            end
+
+            should 'add the percona user to the command' do
+              command = @migration.send(:percona_command, '', '', '')
+              assert command.include?('--user bazqux'), "command #{command} did not include the right user"
+            end
+          end
+
+          context 'with user forbidden in percona config' do
+            setup do
+              @migration.stubs(:raw_database_config).returns(username: 'foobar', percona: { user: false })
+            end
+
+            should 'not add a user to the command' do
+              command = @migration.send(:percona_command, '', '', '')
+              assert !command.include?('--user foobar'), "command #{command} included user but should not have"
+            end
+          end
+        end
+
+        context 'password flag' do
+          context 'with normal database password specified' do
+            setup do
+              @migration.stubs(:raw_database_config).returns(password: 'foobar')
+            end
+
+            should 'add the password to the command' do
+              command = @migration.send(:percona_command, '', '', '')
+              assert command.include?('--password foobar'), "command #{command} did not include password"
+            end
+          end
+
+          context 'with password specified in percona config' do
+            setup do
+              @migration.stubs(:raw_database_config).returns(percona: { password: 'foobar' })
+            end
+
+            should 'add the password to the command' do
+              command = @migration.send(:percona_command, '', '', '')
+              assert command.include?('--password foobar'), "command #{command} did not include password"
+            end
+          end
+
+          context 'with password specified in database and percona config' do
+            setup do
+              @migration.stubs(:raw_database_config).returns(password: 'foobar', percona: { password: 'bazqux' })
+            end
+
+            should 'add the percona password to the command' do
+              command = @migration.send(:percona_command, '', '', '')
+              assert command.include?('--password bazqux'), "command #{command} did not include the right password"
+            end
+          end
+
+          context 'with password forbidden in percona config' do
+            setup do
+              @migration.stubs(:raw_database_config).returns(password: 'foobar', percona: { password: false })
+            end
+
+            should 'not add a password to the command' do
+              command = @migration.send(:percona_command, '', '', '')
+              assert !command.include?('--password foobar'), "command #{command} included password but should not have"
+            end
+          end
+        end
       end
 
       context 'connected to a pt-osc database in print mode' do
@@ -252,6 +344,17 @@ class PtOscMigrationUnitTest < Test::Unit::TestCase
       end
     end
 
+    context '#sanitize_command' do
+      should 'strip password out of commands' do
+        assert_equal 'pt-online-schema-change --password _hidden_', @migration.class.send(:sanitize_command, 'pt-online-schema-change --password supersecret')
+        assert_equal 'pt-online-schema-change --password _hidden_', @migration.class.send(:sanitize_command, 'pt-online-schema-change --password %&$^*(!(#)CD&`+')
+        assert_equal 'pt-online-schema-change --password _hidden_', @migration.class.send(:sanitize_command, 'pt-online-schema-change --password "who uses spaces in passwords?"')
+        assert_equal 'pt-online-schema-change --other --password _hidden_ --fields', @migration.class.send(:sanitize_command, 'pt-online-schema-change --other --password supersecret --fields')
+        assert_equal 'pt-online-schema-change --other --fields --password _hidden_', @migration.class.send(:sanitize_command, 'pt-online-schema-change --other --fields --password supersecret')
+        assert_equal 'pt-online-schema-change --password', @migration.class.send(:sanitize_command, 'pt-online-schema-change --password')
+      end
+    end
+
     context '#make_path_absolute' do
       context 'with an absolute path' do
         setup do
@@ -292,6 +395,28 @@ class PtOscMigrationUnitTest < Test::Unit::TestCase
       end
     end
 
+    context '#get_from_config' do
+      setup do
+        @migration.stubs(:raw_database_config).returns(flag: 'foobar', other_flag: 'mooshu')
+      end
+
+      should 'return nil if false is specified' do
+        assert_nil @migration.send(:get_from_config, false)
+      end
+
+      should 'get the value from the config if it was not specified' do
+        assert_equal 'foobar', @migration.send(:get_from_config, nil, flag_name: 'flag')
+      end
+
+      should 'prefer the key name if given explicitly' do
+        assert_equal 'mooshu', @migration.send(:get_from_config, nil, flag_name: 'flag', key_name: 'other_flag')
+      end
+
+      should 'pass string values through' do
+        assert_equal 'bazqux', @migration.send(:get_from_config, 'bazqux')
+      end
+    end
+
     context '#execute_pt_osc' do
       context 'with a pt-osc connection' do
         setup do
@@ -309,7 +434,7 @@ class PtOscMigrationUnitTest < Test::Unit::TestCase
         context 'connected to a database' do
           setup do
             @database_name = Faker::Lorem.word
-            @migration.stubs(:database_config).returns(database: @database_name)
+            @migration.stubs(:raw_database_config).returns(database: @database_name)
           end
 
           teardown do
@@ -427,9 +552,9 @@ class PtOscMigrationUnitTest < Test::Unit::TestCase
           end
 
           should 'log the command' do
-            @migration.expects(:percona_command).returns('<<percona command>>')
+            @migration.expects(:percona_command).returns('percona command')
             @migration.send(:execute_sql_for_table, nil, nil, nil)
-            assert '<<percona command>>'.in?(@dummy_log.string), 'Log entry did not contain percona command'
+            assert 'percona command'.in?(@dummy_log.string), 'Log entry did not contain percona command'
           end
 
           context 'with successful execution' do